Fixed login and Flight Update

+++ FinalExam/FinalExam_DigiBook/digibook-web/src/app/flights/flights.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { MenuItem, ConfirmationService , DataTable} from 'primeng/primeng';
+import { MenuItem, ConfirmationService, DataTable } from 'primeng/primeng';
 import { Flight } from '../domain/flight';
 import { GlobalService } from '../services/globalservice';
 import { FlightClass } from '../domain/FlightClass';
@@ -7,6 +7,7 @@ import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms'
 import { Pagination } from '../domain/pagination';
 import { MaxLengthValidator } from '@angular/forms/src/directives/validators';
 import { DatePipe } from '@angular/common';
+import { Message } from 'primeng/components/common/api';
 
 @Component({
   selector: 'app-flights',
@@ -31,8 +32,8 @@ export class FlightsComponent implements OnInit {
   flightList: Flight[];
   display: boolean = false;
   userform: FormGroup;
-  today : Date;
-
+  today: Date;
+  msgs: Message[] = [];
   UpperCaseDigitRejex: string = "([A-Z]{3})";
 
   showDialog() {
@@ -40,7 +41,7 @@ export class FlightsComponent implements OnInit {
   }
   constructor(private globalService: GlobalService,
     private confirmationService: ConfirmationService,
-  private fb: FormBuilder, private datePipe: DatePipe) { }
+    private fb: FormBuilder, private datePipe: DatePipe) { }
 
   @ViewChild('dt') public dataTable: DataTable;
   ngOnInit() {
@@ -68,57 +69,68 @@ export class FlightsComponent implements OnInit {
     this.displayDialog = true;
     this.userform.markAsPristine();
   }
-  
-  saveFlight(){
+
+  saveFlight() {
     this.userform.enable();
-    let tmpFlightList = [...this.flightList];
-    if(this.isNewFlight){
-      this.globalService.addSomething("Flights" , this.selectedFlight).then(flights => {
-      this.flight = flights;
-      flights.expectedTimeOfArrival = new Date(flights.expectedTimeOfArrival).toLocaleString();
-      flights.expectedTimeOfDeparture = new Date(flights.expectedTimeOfDeparture).toLocaleString();
-      tmpFlightList.push(this.flight);
-      this.flightList = tmpFlightList;
-      });
-    } else {
-      this.globalService.updateSomething<Flight>("Flights",this.selectedFlight.flightId,this.selectedFlight)
-      .then(flights =>{
-        // this.selectedFlight.expectedTimeOfArrival = new Date(this.selectedFlight.expectedTimeOfArrival).toLocaleString();
-        // this.selectedFlight.expectedTimeOfDeparture = new Date(this.selectedFlight.expectedTimeOfDeparture).toLocaleString();
-        tmpFlightList[this.flightList.indexOf(this.selectedFlight)] = this.selectedFlight;
-        this.flightList=tmpFlightList;
-      });
-    }
-    this.selectedFlight = null;
-    this.isNewFlight = false;
+    this.confirmationService.confirm({
+      message: 'Are you sure that you want to proceed?',
+      header: 'Confirmation',
+      icon: 'fa fa-question-circle',
+      accept: () => {
+        this.msgs = [{ severity: 'info', summary: 'Confirmed', detail: 'You have accepted' }];
+        let tmpFlightList = [...this.flightList];
+        if (this.isNewFlight) {
+          this.globalService.addSomething("Flights", this.selectedFlight).then(flights => {
+            this.flight = flights;
+            flights.expectedTimeOfArrival = new Date(flights.expectedTimeOfArrival).toLocaleString();
+            flights.expectedTimeOfDeparture = new Date(flights.expectedTimeOfDeparture).toLocaleString();
+            tmpFlightList.push(this.flight);
+            this.flightList = tmpFlightList;
+          });
+        } else {
+          this.globalService.updateSomething<Flight>("Flights", this.selectedFlight.flightId, this.selectedFlight)
+            .then(flights => {
+              this.selectedFlight.expectedTimeOfArrival = new Date(this.selectedFlight.expectedTimeOfArrival).toLocaleString();
+              this.selectedFlight.expectedTimeOfDeparture = new Date(this.selectedFlight.expectedTimeOfDeparture).toLocaleString();
+              tmpFlightList[this.flightList.indexOf(this.selectedFlight)] = this.selectedFlight;
+              this.flightList = tmpFlightList;
+            });
+        }
+        this.selectedFlight = null;
+        this.isNewFlight = false;
+      },
+      reject: () => {
+        this.msgs = [{ severity: 'info', summary: 'Rejected', detail: 'You have rejected' }];
+      }
+    });
     this.userform.markAsPristine();
   }
 
-  saveNewFlight(){
+  saveNewFlight() {
     this.userform.enable();
     let tmpFlightList = [...this.flightList];
-    if(this.isNewFlight){
-      this.globalService.addSomething("Flights" , this.selectedFlight).then(flights => {
-      this.flight = flights;
-      flights.expectedTimeOfArrival = new Date(flights.expectedTimeOfArrival).toLocaleString();
-      flights.expectedTimeOfDeparture = new Date(flights.expectedTimeOfDeparture).toLocaleString();
-      tmpFlightList.push(this.flight);
-      this.flightList = tmpFlightList;
+    if (this.isNewFlight) {
+      this.globalService.addSomething("Flights", this.selectedFlight).then(flights => {
+        this.flight = flights;
+        flights.expectedTimeOfArrival = new Date(flights.expectedTimeOfArrival).toLocaleString();
+        flights.expectedTimeOfDeparture = new Date(flights.expectedTimeOfDeparture).toLocaleString();
+        tmpFlightList.push(this.flight);
+        this.flightList = tmpFlightList;
       });
     } else {
-      this.globalService.updateSomething<Flight>("Flights",this.selectedFlight.flightId,this.selectedFlight)
-      .then(flights =>{
-        this.selectedFlight.expectedTimeOfArrival = new Date(this.selectedFlight.expectedTimeOfArrival).toLocaleString();
-        this.selectedFlight.expectedTimeOfDeparture = new Date(this.selectedFlight.expectedTimeOfDeparture).toLocaleString();
-        tmpFlightList[this.flightList.indexOf(this.selectedFlight)] = this.selectedFlight;
-        this.flightList=tmpFlightList;
-      });
+      this.globalService.updateSomething<Flight>("Flights", this.selectedFlight.flightId, this.selectedFlight)
+        .then(flights => {
+          this.selectedFlight.expectedTimeOfArrival = new Date(this.selectedFlight.expectedTimeOfArrival).toLocaleString();
+          this.selectedFlight.expectedTimeOfDeparture = new Date(this.selectedFlight.expectedTimeOfDeparture).toLocaleString();
+          tmpFlightList[this.flightList.indexOf(this.selectedFlight)] = this.selectedFlight;
+          this.flightList = tmpFlightList;
+        });
     }
     this.selectedFlight = new FlightClass;
     this.userform.markAsPristine();
   }
 
-  deleteConfirmation(Flight : Flight){
+  deleteConfirmation(Flight: Flight) {
     this.userform.disable();
     this.btnSaveNew = false;
     this.btnDelete = true;
@@ -133,12 +145,23 @@ export class FlightsComponent implements OnInit {
   findselectedFlightIndex(): number {
     return this.flightList.indexOf(this.selectedFlight);
   }
-  deleteFlight(){
-    let index = this.findselectedFlightIndex();
-    this.flightList = this.flightList.filter((val, i) => i != index);
-    this.globalService.deleteSomething("Flights", this.selectedFlight.flightId);
-    this.selectedFlight = null;
-    this.displayDialog = false;
+  deleteFlight() {
+    this.confirmationService.confirm({
+      message: 'Are you sure that you want to proceed?',
+      header: 'Confirmation',
+      icon: 'fa fa-question-circle',
+      accept: () => {
+        this.msgs = [{ severity: 'info', summary: 'Confirmed', detail: 'You have accepted' }];
+        let index = this.findselectedFlightIndex();
+        this.flightList = this.flightList.filter((val, i) => i != index);
+        this.globalService.deleteSomething("Flights", this.selectedFlight.flightId);
+        this.selectedFlight = null;
+        this.displayDialog = false;
+      },
+      reject: () => {
+        this.msgs = [{ severity: 'info', summary: 'Rejected', detail: 'You have rejected' }];
+      }
+    });
   }
 
   cloneRecord(r: Flight): Flight {
@@ -149,7 +172,7 @@ export class FlightsComponent implements OnInit {
     return flight;
   }
 
-  editFlight(Flight : Flight){
+  editFlight(Flight: Flight) {
     this.userform.enable();
     this.btnSave = false;
     this.btnDelete = false;
@@ -170,12 +193,12 @@ export class FlightsComponent implements OnInit {
         this.paginationResult = paginationResult;
         this.flightList = this.paginationResult.results;
         this.totalRecords = this.paginationResult.totalRecords;
-         for (let i = 0; i < this.flightList.length; i++){
-        this.flightList[i].expectedTimeOfArrival = new Date(this.flightList[i].expectedTimeOfArrival).toLocaleString();
-        this.flightList[i].expectedTimeOfDeparture = new Date(this.flightList[i].expectedTimeOfDeparture).toLocaleString();
-        // this.flightList[i].expectedTimeOfArrival = this.datePipe.transform(this.flightList[i].expectedTimeOfArrival, 'MM-dd-yyyy HH:mm:ss');
-        // this.flightList[i].expectedTimeOfDeparture = this.datePipe.transform(this.flightList[i].expectedTimeOfDeparture, 'MM-dd-yyyy HH:mm:ss');
-         }
+        for (let i = 0; i < this.flightList.length; i++) {
+          this.flightList[i].expectedTimeOfArrival = new Date(this.flightList[i].expectedTimeOfArrival).toLocaleString();
+          this.flightList[i].expectedTimeOfDeparture = new Date(this.flightList[i].expectedTimeOfDeparture).toLocaleString();
+          // this.flightList[i].expectedTimeOfArrival = this.datePipe.transform(this.flightList[i].expectedTimeOfArrival, 'MM-dd-yyyy HH:mm:ss');
+          // this.flightList[i].expectedTimeOfDeparture = this.datePipe.transform(this.flightList[i].expectedTimeOfDeparture, 'MM-dd-yyyy HH:mm:ss');
+        }
       });
   }
 
@@ -193,13 +216,24 @@ export class FlightsComponent implements OnInit {
     };
     this.dataTable.paginate();
   }
-  
-  cancelFlight(){
-    this.isNewFlight = false;
-    let tmpFlightList = [...this.flightList];
-    tmpFlightList[this.flightList.indexOf(this.selectedFlight)] = this.cloneFlight;
-    this.flightList = tmpFlightList;
-    this.selectedFlight = this.cloneFlight;
-    this.selectedFlight = null;
+
+  cancelFlight() {
+    this.confirmationService.confirm({
+      message: 'Are you sure that you want to proceed?',
+      header: 'Confirmation',
+      icon: 'fa fa-question-circle',
+      accept: () => {
+        this.msgs = [{ severity: 'info', summary: 'Confirmed', detail: 'You have accepted' }];
+        this.isNewFlight = false;
+        let tmpFlightList = [...this.flightList];
+        tmpFlightList[this.flightList.indexOf(this.selectedFlight)] = this.cloneFlight;
+        this.flightList = tmpFlightList;
+        this.selectedFlight = this.cloneFlight;
+        this.selectedFlight = null;
+      },
+      reject: () => {
+        this.msgs = [{ severity: 'info', summary: 'Rejected', detail: 'You have rejected' }];
+      }
+    });
   }
 }
